backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
sqlmodel==0.0.22
aiosqlite==0.20.0
openai==1.59.4
faiss-cpu==1.9.0
pymupdf==1.25.1
pdfplumber==0.11.4
python-multipart==0.0.19
tiktoken==0.8.0
spacy==3.8.3
# spaCy model — install separately: python -m spacy download en_core_web_sm
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
# Optional OCR (install manually if needed):
# pytesseract==0.3.13
# pdf2image==1.17.0
